prepare server side functionality for the feature

+++ app/Http/Controllers/ParticipantController.php
@@ -12,8 +12,20 @@ class ParticipantController extends Controller
     {
         $user = Auth::user();
 
-        if ($event->organizer_id === $user->id) {
+        if ($event->organizer_id === $user->id && $event->participants()->find($participant->id)) {
             $event->participants()->detach($participant);
+        }
+
+        return redirect()->route('events.show', $event->id);
+    }
+
+    public function makeOrganizer(Event $event, User $participant)
+    {
+        $user = Auth::user();
+
+        if ($event->organizer_id === $user->id && $event->participants()->find($participant->id)) {
+            $event->organizer_id = $participant->id;
+            $event->save();
         }
 
         return redirect()->route('events.show', $event->id);
